Quick single view on thumbnail page
improvement: single view is a added layer on the thumbnail page now.
This means that no extra loading is required, and the preview is faster
to open and to close. Page rebuilds are not happening anymore.

todo: metadata on single view has not been implemented back, todo:
permalink to single view page